Upate filtering operations
Binary images are now handled. Cleaned up and optimized code. Some speed
gains after shifting to multiplies and not using int8_t.

Added a sharp and unsharp mask feature. Fixed up guassian. Added a
laplacian operation for edge detection.

+++ src/omv/img/fsort.h
@@ -9,5 +9,5 @@
 #ifndef __FSORT_H__
 #define __FSORT_H__
 #include <stdint.h>
-void fsort(uint8_t *data, int n);
+void fsort(int *data, int n);
 #endif /* __FSORT_H__ */
